fix: @Transactional self-injection 문제 해결 - DailyPriceWriter에 업데이트 로직 이전
