+++ src/main/java/org/project/ssogssog/application/service/stock/collect/SyncDailyPriceUseCase.java
@@ -2,6 +2,7 @@ package org.project.ssogssog.application.service.stock.collect;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.project.ssogssog.application.service.stock.writer.DailyPriceWriter;
 import org.project.ssogssog.domain.stock.entity.DailyPrice;
 import org.project.ssogssog.domain.stock.entity.Stock;
 import org.project.ssogssog.domain.stock.repository.DailyPriceRepository;
@@ -24,7 +25,7 @@ import java.util.List;
 public class SyncDailyPriceUseCase {
 
     private final StockRepository stockRepository;
-    private final DailyPriceRepository dailyPriceRepository;
+    private final DailyPriceWriter dailyPriceWriter;
 
     /**
      * 전 종목의 등락 정보 동기화
@@ -40,7 +41,7 @@ public class SyncDailyPriceUseCase {
             int updated = 0;
 
             try {
-                updated = syncDailyPriceChangeInfo(stock);
+                updated = dailyPriceWriter.syncDailyPriceChangeInfo(stock);
                 totalUpdated += updated;
             } catch (Exception e) {
                 log.error("등락 정보 동기화 실패 - 종목: {}({}), 에러: {}",
@@ -54,51 +55,6 @@ public class SyncDailyPriceUseCase {
         }
 
         log.info("등락 정보 동기화 완료 - 총 {}개 종목, {}개 데이터 업데이트", stocks.size(), totalUpdated);
-    }
-
-    /**
-     * 특정 종목의 일별시세 등락 정보 동기화
-     * N+1 문제 방지: 해당 종목의 모든 DailyPrice를 한번에 조회
-     */
-    @Transactional
-    public int syncDailyPriceChangeInfo(Stock stock) {
-        // 한번에 모든 DailyPrice 조회 (날짜 오름차순)
-        List<DailyPrice> dailyPrices = dailyPriceRepository.findByStockOrderByDateAsc(stock);
-
-        if (dailyPrices.isEmpty()) {
-            return 0;
-        }
-
-        List<DailyPrice> updatedPrices = calculateChangeInfo(dailyPrices);
-
-        if (!updatedPrices.isEmpty()) {
-            dailyPriceRepository.saveAll(updatedPrices);
-            log.debug("종목 {} - {}개 데이터 등락 정보 업데이트", stock.getStockCode(), updatedPrices.size());
-        }
-
-        return updatedPrices.size();
-    }
-
-    /**
-     * 연속된 DailyPrice 리스트에서 등락 정보가 없는 데이터를 계산
-     * 계산 로직은 DailyPrice 엔티티에 위임
-     *
-     * @param dailyPrices 날짜순 정렬된 DailyPrice 리스트
-     * @return 업데이트된 DailyPrice 리스트
-     */
-    private List<DailyPrice> calculateChangeInfo(List<DailyPrice> dailyPrices) {
-        List<DailyPrice> updated = new ArrayList<>();
-        DailyPrice prev = null;
-
-        for (DailyPrice current : dailyPrices) {
-            if (current.needsChangeInfoSync() && prev != null) {
-                current.syncChangeInfo(prev.getClosePrice());
-                updated.add(current);
-            }
-            prev = current;
-        }
-
-        return updated;
     }
 
 }
